test(api): move error acceptance tests to integration
