+++ api/tests/team/acceptance/application/certification-center-membership/certification-center-membership.route.test.js
@@ -44,27 +44,21 @@ describe('Acceptance | Team | Application | Routes | certification-center-member
   });
 
   describe('PATCH /api/certification-centers/{certificationCenterId}/certification-center-memberships/{id}', function () {
-    let certificationCenter;
-    let certificationCenterMembership;
-    let user;
-
-    beforeEach(async function () {
-      certificationCenter = databaseBuilder.factory.buildCertificationCenter();
-      user = databaseBuilder.factory.buildUser();
-      certificationCenterMembership = databaseBuilder.factory.buildCertificationCenterMembership({
-        certificationCenterId: certificationCenter.id,
-        userId: user.id,
-      });
-      await databaseBuilder.commit();
-    });
-
     context('Success cases', function () {
       it('returns a 200 HTTP status code with the updated certification center membership', async function () {
         // given
+        const certificationCenter = databaseBuilder.factory.buildCertificationCenter();
+        const user = databaseBuilder.factory.buildUser();
+        const certificationCenterMembership = databaseBuilder.factory.buildCertificationCenterMembership({
+          certificationCenterId: certificationCenter.id,
+          userId: user.id,
+        });
         const certifCenterAdminUser = databaseBuilder.factory.buildUser.withCertificationCenterMembership({
           certificationCenterId: certificationCenter.id,
           role: 'ADMIN',
         });
+        await databaseBuilder.commit();
+
         const request = {
           method: 'PATCH',
           url: `/api/certification-centers/${certificationCenter.id}/certification-center-memberships/${certificationCenterMembership.id}`,
@@ -80,7 +74,6 @@ describe('Acceptance | Team | Application | Routes | certification-center-member
           },
           headers: generateAuthenticatedUserRequestHeaders({ userId: certifCenterAdminUser.id }),
         };
-        await databaseBuilder.commit();
 
         // when
         const response = await server.inject(request);
@@ -101,113 +94,6 @@ describe('Acceptance | Team | Application | Routes | certification-center-member
           },
         };
         expect(_.omit(response.result, 'included')).to.deep.equal(expectedUpdatedCertificationCenterMembership);
-      });
-    });
-
-    context('Error cases', function () {
-      context('when current user has a member role', function () {
-        it('returns a 403 HTTP status code', async function () {
-          // given
-          const certifCenterMemberUser = databaseBuilder.factory.buildUser.withCertificationCenterMembership({
-            certificationCenterId: certificationCenter.id,
-            role: 'MEMBER',
-          });
-          const request = {
-            method: 'PATCH',
-            url: `/api/certification-centers/${certificationCenter.id}/certification-center-memberships/${certificationCenterMembership.id}`,
-            payload: {
-              id: user.id,
-              data: {
-                type: 'certification-center-memberships',
-                'certification-center-membership-id': certificationCenterMembership.id.toString(),
-                attributes: {
-                  role: 'ADMIN',
-                },
-              },
-            },
-            headers: generateAuthenticatedUserRequestHeaders({ userId: certifCenterMemberUser.id }),
-          };
-          await databaseBuilder.commit();
-
-          // when
-          const { statusCode } = await server.inject(request);
-
-          // then
-          expect(statusCode).to.equal(403);
-        });
-      });
-
-      context('when certification center membership does not belong to the certification center', function () {
-        it('returns a 403 HTTP status code', async function () {
-          // given
-          const certifCenterAdminUser = databaseBuilder.factory.buildUser.withCertificationCenterMembership({
-            certificationCenterId: certificationCenter.id,
-            role: 'ADMIN',
-          });
-
-          const anotherCertificationCenter = databaseBuilder.factory.buildCertificationCenter();
-          const anotherCertificationCenterMembership = databaseBuilder.factory.buildCertificationCenterMembership({
-            certificationCenterId: anotherCertificationCenter.id,
-          });
-          const anotherCertificationCenterMembershipId = anotherCertificationCenterMembership.id;
-
-          const request = {
-            method: 'PATCH',
-            url: `/api/certification-centers/${certificationCenter.id}/certification-center-memberships/${anotherCertificationCenterMembershipId}`,
-            payload: {
-              data: {
-                id: user.id,
-                type: 'certification-center-memberships',
-                attributes: {
-                  role: 'ADMIN',
-                  'certification-center-membership-id': anotherCertificationCenterMembershipId,
-                },
-              },
-            },
-            headers: generateAuthenticatedUserRequestHeaders({ userId: certifCenterAdminUser.id }),
-          };
-          await databaseBuilder.commit();
-
-          // when
-          const { statusCode } = await server.inject(request);
-
-          // then
-          expect(statusCode).to.equal(403);
-        });
-      });
-
-      context('when certification center membership ID in url is not valid', function () {
-        it('returns a 400 HTTP status code', async function () {
-          // given
-          const certifCenterAdminUser = databaseBuilder.factory.buildUser.withCertificationCenterMembership({
-            certificationCenterId: certificationCenter.id,
-            role: 'ADMIN',
-          });
-          const wrongCertificationCenterMembershipId = certificationCenterMembership.id + 1;
-          const notValidMembershipId = `toto`;
-          const request = {
-            method: 'PATCH',
-            url: `/api/certification-centers/${certificationCenter.id}/certification-center-memberships/${notValidMembershipId}`,
-            payload: {
-              data: {
-                id: user.id,
-                type: 'certification-center-memberships',
-                attributes: {
-                  role: 'ADMIN',
-                  'certification-center-membership-id': wrongCertificationCenterMembershipId,
-                },
-              },
-            },
-            headers: generateAuthenticatedUserRequestHeaders({ userId: certifCenterAdminUser.id }),
-          };
-          await databaseBuilder.commit();
-
-          // when
-          const { statusCode } = await server.inject(request);
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
       });
     });
   });
